+++ src/Servidor/Modelo/personaje.cpp
@@ -35,7 +35,7 @@ void Personaje::notificar_observadores(){
 	}
 }
 
-void Personaje::update(size_t tiempo, Mapa* mapa){
+void Personaje::update(float tiempo, Mapa* mapa){
 	tiempo_pasado += tiempo;
 	mover(tiempo, mapa);
 	notificar_observadores();
